+++ example/lib/multiple_view_models_example/view_model/first_view_model.dart
@@ -16,6 +16,6 @@ class FirstViewModel extends ViewModel {
   void dispose() {
     // must dispose all flows
     _counterStateFlow.dispose();
-    debugPrint("ViewModelA disposed");
+    debugPrint("FirstViewModel disposed");
   }
 }