+++ example/lib/multiple_view_models_example/view_model/second_view_model.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:view_model_x/view_model_x.dart';
 
 class SecondViewModel extends ViewModel with PostFrameCallback {
@@ -8,16 +9,23 @@ class SecondViewModel extends ViewModel with PostFrameCallback {
 
   void showPopupMessage() {
     // by emitting the value, listeners were notified
+    debugPrint("hi");
     _messageSharedFlow.emit("Hello from MyViewModel!");
   }
 
   @override
-  void dispose() {
-    _messageSharedFlow.dispose();
+  void init() {
+    debugPrint("init inside vm");
   }
 
   @override
   void onPostFrameCallback(Duration timestamp) {
     _messageSharedFlow.emit("onPostFrameCallback from SecondViewModel");
+  }
+
+  @override
+  void dispose() {
+    _messageSharedFlow.dispose();
+    debugPrint("FirstViewModel disposed");
   }
 }