+++ lib/provider.dart
@@ -1,0 +1,2 @@
+export 'package:provider/provider.dart' show ProviderNotFoundException;
+export './src/modified_provider.dart';