+++ lib/src/provider_single_child_widget.dart
@@ -1,0 +1,5 @@
+import 'package:provider/single_child_widget.dart';
+
+/// Mixin which allows `MultiViewModelProvider` to infer the types
+/// of multiple [ViewModelProvider]s.
+mixin ProviderSingleChildWidget on SingleChildWidget {}