+++ lib/view_model_x.dart
@@ -6,6 +6,5 @@ export './src/state_flow_builder.dart';
 export './src/state_flow_consumer.dart';
 export './src/state_flow_listener.dart';
 export './src/shared_flow_listener.dart';
-export './src/view_model_provider.dart' hide ViewModelProviderSingleChildWidget;
-export './src/multi_view_model_provider.dart';
+export './src/view_model_provider.dart';
 export './src/multi_flow_listener.dart';